requirements.txt
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.4.0
autocommand==2.2.2
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.1.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.3
crit==4.2
cryptography==45.0.7
cssselect==1.3.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.4.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.2
fonttools==4.57.0
fs==2.4.16
gyp-next==0.16.2
html5lib-modern==1.2
httplib2==0.22.0
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
kali-tweaks==2025.4.0
kazam==1.4.5
kiwisolver==1.4.10rc0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
legacy-cgi==2.6.4
lightdm-gtk-greeter-settings==1.2.3
llvmlite==0.44.0
louis==3.35.0
lxml==6.0.2
lz4==4.4.4+dfsg
MarkupSafe==3.0.3
matplotlib==3.10.7+dfsg1
mechanize==0.4.10
more-itertools==10.8.0
mpmath==1.3.0
netifaces==0.11.0
numba==0.61.2
numexpr==2.14.1
numpy==2.3.5
oauthlib==3.3.1
odfpy==1.4.2
olefile==0.47
onboard==1.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
perf==0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==4.21.12
psutil==7.1.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasyncore==1.0.2
pycriu==4.2
pycups==2.0.4
Pygments==2.18.0
PyGObject==3.50.2
pyinotify==0.9.6
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.1.3
PyQt6==6.10.0
PyQt6_sip==13.10.2
pysmbc==1.0.25.1
python-apt==3.0.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-linux-procfs==0.7.3
python-magic==0.4.27
pytz==2025.2
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.2
requests==2.32.5
scipy==1.16.3
seaborn==0.13.2
setproctitle==1.3.7
soupsieve==2.7
sympy==1.14.0
tables==3.10.2
tqdm==4.67.1
typeguard==4.4.4
typing_extensions==4.15.0
ufoLib2==0.17.1
unicodedata2==16.0.0
urllib3==2.5.0
wadllib==2.0.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.46.1
xdg==5
zopfli==0.4.0
